// FordTube.Web/src/app/domain/interfaces/slide.interface.ts
/**
  * Model/Interface for: FordTube.VBrick.Wrapper.Models.SlideModel
  */
export interface SlideModel {

  id: number;
  activeDate?: Date;
  activeTime?: string;
  inactiveDate?: Date;
  inactiveTime?: string;
  franchise: number;
  backgroundImageUrl: string;
  backgroundColor: string;
  link: string;
  text: string;
  textPosition: number;
}