// FordTube.Web/src/app/domain/interfaces/slidedto.interface.ts
/**
  * Model/Interface for: FordTube.VBrick.Wrapper.Models.SlideDtoModel
  */
export interface SlideDtoModel {

  id: number;
  activeDate?: Date;
  activeTime?: string;
  inactiveDate?: Date;
  inactiveTime?: string;
  franchise: number;
  backgroundImage: number[];
  backgroundImageUrl: string;
  backgroundColor: string;
  link: string;
  text: string;
  textPosition: number;
}